+++ airflute_v2/src/combinationToMidi.h
@@ -28,6 +28,6 @@
 #include <stdint.h>
 
 uint8_t combinationToMidi(uint8_t sensorValues);
-int hammingDistance(uint8_t a, uint8_t b);
+uint8_t hammingDistance(uint8_t a, uint8_t b);
 
 #endif
